# taskboard/public/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TaskBoard API",
    "version": "1.0.0",
    "description": "A comprehensive API for managing Kanban tasks with integrated MCP (Model Context Protocol) chat functionality.\n\n## Authentication\n\nProtected endpoints require user identification via one of these methods:\n1. **Bearer Token**: `Authorization: Bearer <userId>`\n2. **Query Parameter**: `?userId=<userId>`  \n3. **Request Body**: Include `userId` field in JSON body\n\nThis flexible authentication supports both web applications and MCP integrations.\n\n## Features\n\n- User registration and authentication\n- CRUD operations for Kanban items (todo, doing, done)  \n- MCP-powered AI chat assistant integration\n- Session-based conversation management\n- Tool usage tracking and display"
  },
  "servers": [
    {
      "url": "${API_BASE_URL:-http://localhost:3000}",
      "description": "API server (configurable via API_BASE_URL environment variable)"
    }
  ],
...
  "paths": {
...
    "/api/items": {
      "get": {
        "summary": "Get user's items",
        "description": "Retrieve all items belonging to the authenticated user, optionally filtered by status",
        "operationId": "get_user_items",
        "tags": [
          "Items"
        ],
        "x-gram": {
          "name": "get_user_tasks",
          "description": "<context>\nThis endpoint retrieves all Kanban tasks belonging to the authenticated user.\nTasks can be filtered by status (todo, doing, done) to focus on specific work stages.\n</context>\n<prerequisites>\n- userId parameter is required for authentication\n</prerequisites>\n<usage>\n- Use without status parameter to get all tasks\n- Use status=todo to get pending tasks\n- Use status=doing to get in-progress tasks  \n- Use status=done to get completed tasks\n</usage>\n",
          "responseFilterType": "jq"
        },
        "security": [
          {
            "UserIdParam": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter items by status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ItemStatus"
            },
            "example": "todo"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "User ID (required for authentication)",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "clh7x1234abcd5678user"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                },
                "example": [
                  {
                    "id": "clh7x1234abcd5678efgh",
                    "userId": "clh7x1234abcd5678user",
                    "title": "Complete project documentation",
                    "details": "Write comprehensive API documentation and user guides",
                    "status": "todo",
                    "createdAt": "2025-07-31T20:30:00.000Z"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Unauthorized"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new item",
        "description": "Create a new Kanban item for the authenticated user",
        "operationId": "create_user_item",
        "tags": [
          "Items"
        ],
        "x-gram": {
          "name": "create_task",
          "description": "<context>\nThis endpoint creates a new Kanban task for the authenticated user.\nTasks can be created in any status (todo, doing, done) and will be associated with the user's account.\n</context>\n<prerequisites>\n- userId parameter is required for authentication\n- Title, details, and status are required fields\n</prerequisites>\n<usage>\n- Use status=todo for new tasks that need to be started\n- Use status=doing for tasks currently in progress\n- Use status=done for completed tasks\n- Title should be concise (max 200 characters)\n- Details can include full description (max 1000 characters)\n</usage>\n",
          "responseFilterType": "jq"
        },
        "security": [
          {
            "UserIdParam": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemRequest"
              },
              "example": {
                "title": "Implement user authentication",
                "details": "Add JWT-based authentication with refresh tokens",
                "status": "todo",
                "userId": "clh7x1234abcd5678user"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                },
                "example": {
                  "id": "clh7x9012qrst3456uvwx",
                  "userId": "clh7x1234abcd5678user",
                  "title": "Implement user authentication",
                  "details": "Add JWT-based authentication with refresh tokens",
                  "status": "todo",
                  "createdAt": "2025-07-31T21:45:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Missing title, details, or status"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Unauthorized"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ItemStatus": {
        "type": "string",
        "enum": [
          "todo",
          "doing",
          "done"
        ],
        "description": "Current status of the item in the Kanban board"
      },
      "Item": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "title",
          "details",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the item"
          },
          "userId": {
            "type": "string",
            "description": "ID of the user who owns this item"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "description": "Title of the item"
          },
          "details": {
            "type": "string",
            "maxLength": 1000,
            "description": "Detailed description"
          },
          "status": {
            "$ref": "#/components/schemas/ItemStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateItemRequest": {
        "type": "object",
        "required": [
          "title",
          "details",
          "status"
        ],
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "details": {
            "type": "string",
            "maxLength": 1000
          },
          "status": {
            "$ref": "#/components/schemas/ItemStatus"
          },
          "userId": {
            "type": "string",
            "description": "User ID (optional if provided via Bearer token or query param)"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      }
    },
    "securitySchemes": {
      "UserIdParam": {
        "type": "apiKey",
        "in": "query",
        "name": "userId",
        "description": "User ID parameter for authentication"
      }
    }
  }
}